fix: split lab command contract surface
